requirements.txt
asgiref==3.5.2
asttokens
async-generator==1.10
attrs==22.1.0
backcall
backports.functools-lru-cache
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
cchardet==2.1.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama
debugpy
decorator
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.1.1
django-dbbackup==4.0.1
django-guardian==2.4.0
django-haystack==3.2.1
django-import-export==2.8.0
django-multiselectfield==0.1.12
django-parler==2.3
django-rosetta==0.9.8
django-summernote==0.8.20.0
entrypoints
et-xmlfile==1.1.0
executing
gunicorn==20.0.4
h11==0.13.0
idna==3.3
ipykernel
ipython
jedi
jieba==0.42.1
jupyter_client
jupyter_core
lxml==4.9.1
MarkupPy==1.14
matplotlib-inline
nest-asyncio
numpy==1.22.4
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging
pandas==1.4.2
parso
pickleshare
Pillow==9.0.1
polib==1.1.1
prompt-toolkit==3.0.31
psutil
psycopg2==2.8.6
psycopg2-binary==2.8.6
pure-eval
pycparser==2.21
Pygments
pyparsing
PySocks==1.7.1
python-dateutil
python-gettext==4.1
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq
requests==2.28.1
selenium==4.4.3
six
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data
tablib==3.2.1
tornado
tqdm==4.64.1
traitlets
trio==0.21.0
trio-websocket==0.9.2
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.10
wcwidth
webencodings==0.5.1
whitenoise==5.2.0
Whoosh==2.7.4
wincertstore==0.2
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
